refactor(dashboard): split theme constants into server-safe module (AEL-45)
CodeRabbit round 2: keep `lib/theme.ts` as the `"use client"` hook
surface and move the constants the Server Component layout actually
imports (`DEFAULT_THEME`, `THEME_STORAGE_KEY`, `THEME_INIT_SCRIPT`,
`Theme` type) into a new `lib/theme-tokens.ts` module with no client
directive.

This avoids dragging the layout across a `"use client"` boundary just
to read a few strings, and keeps the client-only subscriber set + hook
isolated to the file that owns React-specific code.

- lib/theme-tokens.ts: new server-safe module, exports types + constants
  + the inline pre-paint script body
- lib/theme.ts: re-anchored on theme-tokens; still owns the
  `useSyncExternalStore` store and the `useTheme()` hook
- app/layout.tsx: imports DEFAULT_THEME / THEME_INIT_SCRIPT from the
  server-safe module
- __tests__/lib/theme.test.ts: imports constants from theme-tokens,
  hook from theme

Build clean; 121 unit tests pass.

Made-with: Cursor
