+++ products/dashboard/lib/theme.ts
@@ -1,41 +1,28 @@
 "use client";
 
 /**
- * Theme runtime helpers.
+ * Theme runtime — client-only surface.
  *
- * The dashboard ships dark-first: server renders `<html data-theme="dark">`
- * and an inline pre-paint script in `app/layout.tsx` rehydrates the user's
- * persisted preference before first paint to avoid a theme flash. This
- * module is the canonical client-side surface for reading and writing that
- * preference.
+ * The dashboard ships dark-first: the server renders
+ * `<html data-theme="dark">` and an inline pre-paint script in
+ * `app/layout.tsx` rehydrates the user's persisted preference before first
+ * paint to avoid a theme flash. This module owns the client-side hook.
+ *
+ * Server Components that only need the constants or the inline init
+ * script must import from `lib/theme-tokens.ts` instead, so they do not
+ * cross the client boundary that `"use client"` defines here.
  *
  * State is centralised in a module-level subscriber set so every
  * `useTheme()` consumer (current or future) sees the same value. Without
  * that, two components calling the hook would each keep an independent
- * `useState` snapshot, drift after the first toggle, and silently flip the
- * wrong way on the next call. `useSyncExternalStore` keeps the React
- * renders coherent across all subscribers.
+ * snapshot, drift after the first toggle, and silently flip the wrong way
+ * on the next call. `useSyncExternalStore` keeps the React renders
+ * coherent across all subscribers.
  */
 
 import { useCallback, useSyncExternalStore } from "react";
 
-export type Theme = "dark" | "light";
-
-export const DEFAULT_THEME: Theme = "dark";
-
-/** localStorage key used to persist the user's theme preference. */
-export const THEME_STORAGE_KEY = "theme";
-
-/** Inline script body used by `RootLayout` to rehydrate the persisted theme before paint. */
-export const THEME_INIT_SCRIPT = `(() => {
-  try {
-    const stored = window.localStorage.getItem(${JSON.stringify(THEME_STORAGE_KEY)});
-    const theme = stored === "light" || stored === "dark" ? stored : ${JSON.stringify(DEFAULT_THEME)};
-    document.documentElement.setAttribute("data-theme", theme);
-  } catch {
-    document.documentElement.setAttribute("data-theme", ${JSON.stringify(DEFAULT_THEME)});
-  }
-})();`;
+import { DEFAULT_THEME, THEME_STORAGE_KEY, type Theme } from "./theme-tokens";
 
 function isTheme(value: unknown): value is Theme {
   return value === "dark" || value === "light";
